gcm_ghash_v8: Remove swap of H-table values -- 1st step
This is the first in a series of commits aiming to rewrite gcm_ghash_v8
to work directly with the swapped H-table values, rather than swapping them
back after loading and falling back to the old code.

As a first step, the swapping of A = {H,H2} are removed and all uses of

```
pmull.64 Y, A, X
```

replaced by the equivalent

```
vext.8 X, X, X, #8
pmull2.64 Y, A, X
vext.8 X, X, X, #8
```

(and similarly for pmull2).

This works so long as X and Y don't alias.

Of course, the above conversion makes the code much less efficient,
and is not final. The next commit will eliminate `vext`.