gcm_ghash_v8_4x: Remove initial H-table swaps
As previously done in many small-step commits for gcm_ghash_v8, this commit
modifies gcm_ghash_v8_4x to not swap the H-table values immediately after
loading. The code already juggles the original and shifted form, and it's
a matter of carefully tracking swaps and changing vmullp <-> vmullp2 as
necessary.